Update DOI rules and implementation, abstract regex check
