+++ web/src/views/SubmissionPortal/Components/MultiOmicsDataForm.vue
@@ -1,6 +1,4 @@
 <script lang="ts">
-// @ts-ignore
-import NmdcSchema from 'nmdc-schema/nmdc_schema/nmdc_materialized_patterns.yaml';
 import {
   defineComponent,
   ref,
@@ -10,9 +8,10 @@ import {
 } from '@vue/composition-api';
 
 import Definitions from '@/definitions';
+import doiProviderValues from '@/schema';
 import {
   multiOmicsForm, multiOmicsFormValid, multiOmicsAssociations, checkJGITemplates, canEditSubmissionMetadata, addAwardDoi, removeAwardDoi,
-  templateHasData,
+  templateHasData, checkDoiFormat,
 } from '../store';
 import { AwardTypes, HARMONIZER_TEMPLATES } from '@/views/SubmissionPortal/types';
 
@@ -32,13 +31,6 @@ export default defineComponent({
   },
   setup() {
     const formRef = ref();
-
-    const doiProviderValues = Object.values(NmdcSchema.enums.DoiProviderEnum.permissible_values)
-      .map((pv: any) => ({
-        value: pv.text,
-        text: pv.title || pv.text, // The permissible value definition should have a title, but in case it doesn't fallback to text (which it will always have)
-      }))
-      .sort((a, b) => a.text.localeCompare(b.text));
 
     function reValidate() {
       formRef.value.validate();
@@ -70,13 +62,16 @@ export default defineComponent({
       }
       return true;
     }];
+
     const doiRequiredRules = () => [(v: string) => {
       const valid = !!v || (!multiOmicsForm.facilityGenerated && multiOmicsForm.dataGenerated) || (multiOmicsForm.unknownDoi && !multiOmicsForm.dataGenerated);
-      if (!/^doi:10.\d{2,9}.*$/.test(v)) {
-        return 'Doi must be in the correct format.';
-      }
       return valid || 'An award DOI is required if data was or is going to be generated by a DOE facility.';
+    },
+    (v: string) => {
+      const valid = checkDoiFormat(v);
+      return valid || 'DOI must be in the correct format.';
     }];
+
     const revalidate = () => {
       nextTick(() => formRef.value.validate());
     };
@@ -383,7 +378,7 @@ export default defineComponent({
                 dense
                 clearable
                 class="mb-2 mr-3"
-                :error-messages="multiOmicsForm.awardDois[i].provider ? undefined : ['A provider must be selected.']"
+                :rules="multiOmicsForm.awardDois[i].provider ? undefined : ['A provider must be selected.']"
               >
                 <template #message="{ message }">
                   <span v-html="message" />
@@ -412,7 +407,7 @@ export default defineComponent({
         />
       </div>
       <v-btn
-        v-if="multiOmicsForm.facilities.includes('EMSL') || multiOmicsForm.facilities.includes('JGI')"
+        v-if="multiOmicsForm.dataGenerated || multiOmicsForm.facilities.includes('EMSL') || multiOmicsForm.facilities.includes('JGI')"
         class="mb-4"
         depressed
         @click="addAwardDoi"
